requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
backcall==0.2.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==3.0.7
dominate==2.5.1
future==0.18.2
gast==0.2.2
google-auth==1.17.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
ipython==7.15.0
ipython-genutils==0.2.0
ISR==2.2.0
jedi==0.17.0
jsonpatch==1.25
jsonpointer==2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
protobuf==3.12.2
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.1
scipy==1.2.1
six==1.15.0
sqlparse==0.3.1
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
torch==1.5.0+cu101
torchfile==0.1.0
torchvision==0.6.0+cu101
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
urllib3==1.25.9
visdom==0.1.8.9
wcwidth==0.2.4
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
